обновил мозг милишника, подключив конфигурацию движения

+++ Assets/TestsEnv/RakEnv/_Scripts/Entities/Agents/_Core/IMeleeMotionConfig.cs
@@ -3,8 +3,8 @@
     public interface IMeleeMotionConfig
     {
         float RunSpeed { get; }
-        int RunSpeedZombie { get; }
+        float RunSpeedZombie { get; }
         float WalkSpeed { get; }
-        int WalkSpeedZombie { get; }
+        float WalkSpeedZombie { get; }
     }
 }
